requirements.txt
autopep8
python-dotenv
langchain
openai
streamlit